backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
websockets==13.1
pydantic==2.9.2
supabase==2.10.0
httpx==0.27.2
joblib==1.4.2
scikit-learn==1.5.2
numpy==1.26.4
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
